+++ source/renderer/app/stores/ProfileStore.js
@@ -2,7 +2,7 @@
 import { action, observable, computed, toJS, runInAction } from 'mobx';
 import BigNumber from 'bignumber.js';
 import moment from 'moment/moment';
-import { includes } from 'lodash';
+import { includes, camelCase } from 'lodash';
 import Store from './lib/Store';
 import Request from './lib/LocalizedRequest';
 import { THEMES } from '../themes/index';
@@ -15,7 +15,6 @@ import { Logger } from '../utils/logging';
 import { setStateSnapshotLogChannel } from '../ipc/setStateSnapshotLogChannel';
 import { detectSystemLocaleChannel } from '../ipc/detect-system-locale';
 import { LOCALES } from '../../../common/types/locales.types';
-import { LANGUAGE_OPTIONS } from '../config/profileConfig';
 import {
   compressLogsChannel,
   downloadLogsChannel,
@@ -23,22 +22,33 @@ import {
 } from '../ipc/logs.ipc';
 import type { LogFiles, CompressedLogStatus } from '../types/LogTypes';
 import type { StateSnapshotLogParams } from '../../../common/types/logging.types';
+import {
+  DEFAULT_NUMBER_FORMAT,
+  NUMBER_FORMATS,
+} from '../../../common/types/number.types';
+import {
+  hasLoadedRequest,
+  isRequestSet,
+  requestGetter,
+  getRequestKeys,
+} from '../utils/storesUtils';
+import {
+  NUMBER_OPTIONS,
+  DATE_ENGLISH_OPTIONS,
+  DATE_JAPANESE_OPTIONS,
+  TIME_OPTIONS,
+  PROFILE_SETTINGS,
+} from '../config/profileConfig';
 
 // TODO: refactor all parts that rely on this to ipc channels!
 const { ipcRenderer } = global;
 
 export default class ProfileStore extends Store {
-  LANGUAGE_OPTIONS = LANGUAGE_OPTIONS;
-
   @observable systemLocale: string = LOCALES.english;
-  @observable bigNumberDecimalFormat = {
-    decimalSeparator: '.',
-    groupSeparator: ',',
-    groupSize: 3,
-    secondaryGroupSize: 0,
-    fractionGroupSeparator: ' ',
-    fractionGroupSize: 0,
-  };
+  @observable systemNumberFormat: string = NUMBER_OPTIONS[0].value;
+  @observable systemDateFormatEnglish: string = DATE_ENGLISH_OPTIONS[0].value;
+  @observable systemDateFormatJapanese: string = DATE_JAPANESE_OPTIONS[0].value;
+  @observable systemTimeFormat: string = TIME_OPTIONS[0].value;
 
   /* eslint-disable max-len */
   @observable getProfileLocaleRequest: Request<string> = new Request(
@@ -46,6 +56,32 @@ export default class ProfileStore extends Store {
   );
   @observable setProfileLocaleRequest: Request<string> = new Request(
     this.api.localStorage.setUserLocale
+  );
+  @observable getProfileNumberFormatRequest: Request<string> = new Request(
+    this.api.localStorage.getUserNumberFormat
+  );
+  @observable setProfileNumberFormatRequest: Request<string> = new Request(
+    this.api.localStorage.setUserNumberFormat
+  );
+  @observable getProfileDateFormatEnglishRequest: Request<string> = new Request(
+    this.api.localStorage.getUserDateFormatEnglish
+  );
+  @observable setProfileDateFormatEnglishRequest: Request<string> = new Request(
+    this.api.localStorage.setUserDateFormatEnglish
+  );
+  @observable
+  getProfileDateFormatJapaneseRequest: Request<string> = new Request(
+    this.api.localStorage.getUserDateFormatJapanese
+  );
+  @observable
+  setProfileDateFormatJapaneseRequest: Request<string> = new Request(
+    this.api.localStorage.setUserDateFormatJapanese
+  );
+  @observable getProfileTimeFormatRequest: Request<string> = new Request(
+    this.api.localStorage.getUserTimeFormat
+  );
+  @observable setProfileTimeFormatRequest: Request<string> = new Request(
+    this.api.localStorage.setUserTimeFormat
   );
   @observable getTermsOfUseAcceptanceRequest: Request<string> = new Request(
     this.api.localStorage.getTermsOfUseAcceptance
@@ -72,10 +108,16 @@ export default class ProfileStore extends Store {
   @observable compressedLogsFilePath: ?string = null;
   @observable compressedLogsStatus: CompressedLogStatus = {};
   @observable isSubmittingBugReport: boolean = false;
+  @observable isInitialScreen: boolean = false;
   /* eslint-enable max-len */
 
   setup() {
-    this.actions.profile.updateLocale.listen(this._updateLocale);
+    this.actions.profile.finishInitialScreenSettings.listen(
+      this._finishInitialScreenSettings
+    );
+    this.actions.profile.updateUserLocalSetting.listen(
+      this._updateUserLocalSetting
+    );
     this.actions.profile.acceptTermsOfUse.listen(this._acceptTermsOfUse);
     this.actions.profile.acceptDataLayerMigration.listen(
       this._acceptDataLayerMigration
@@ -90,60 +132,91 @@ export default class ProfileStore extends Store {
     this.actions.app.initAppEnvironment.listen(() => {});
 
     this.registerReactions([
-      this._setBigNumberFormat,
-      this._updateMomentJsLocaleAfterLocaleChange,
-      this._reloadAboutWindowOnLocaleChange,
-      this._redirectToLanguageSelectionIfNoLocaleSet,
+      this._updateBigNumberFormat,
+      // this._updateMomentJsLocaleAfterLocaleChange,
+      // this._reloadAboutWindowOnLocaleChange,
+      this._redirectToInitialSettingsIfNoLocaleSet,
       this._redirectToTermsOfUseScreenIfTermsNotAccepted,
       this._redirectToDataLayerMigrationScreenIfMigrationHasNotAccepted,
       this._redirectToMainUiAfterTermsAreAccepted,
       this._redirectToMainUiAfterDataLayerMigrationIsAccepted,
     ]);
-    this._getSystemLocale();
+    // this._getSystemLocale();
     this._getTermsOfUseAcceptance();
     this._getDataLayerMigrationAcceptance();
   }
 
-  _setBigNumberFormat = () => {
-    BigNumber.config({ FORMAT: this.bigNumberDecimalFormat });
+  _updateBigNumberFormat = () => {
+    const FORMAT = {
+      ...DEFAULT_NUMBER_FORMAT,
+      ...NUMBER_FORMATS[this.currentNumberFormat],
+    };
+    BigNumber.config({ FORMAT });
   };
 
   @computed get currentLocale(): string {
-    const { result } = this.getProfileLocaleRequest.execute();
-    if (this.isCurrentLocaleSet) return result;
-    return this.systemLocale;
+    return requestGetter(this.getProfileLocaleRequest, this.systemLocale);
   }
 
   @computed get hasLoadedCurrentLocale(): boolean {
-    return (
-      this.getProfileLocaleRequest.wasExecuted &&
-      this.getProfileLocaleRequest.result !== null
-    );
+    return hasLoadedRequest(this.getProfileLocaleRequest);
   }
 
   @computed get isCurrentLocaleSet(): boolean {
-    return (
-      this.getProfileLocaleRequest.result !== null &&
-      this.getProfileLocaleRequest.result !== ''
-    );
+    return isRequestSet(this.getProfileLocaleRequest);
   }
 
   @computed get currentTheme(): string {
-    const { result } = this.getThemeRequest.execute();
-    if (this.isCurrentThemeSet) return result;
-    return this.environment.isMainnet ? THEMES.DARK_BLUE : THEMES.LIGHT_BLUE; // defaults
+    const systemValue = this.environment.isMainnet
+      ? THEMES.DARK_BLUE
+      : THEMES.LIGHT_BLUE; // defaults
+    return requestGetter(this.getThemeRequest, systemValue);
   }
 
   @computed get isCurrentThemeSet(): boolean {
-    return (
-      this.getThemeRequest.result !== null && this.getThemeRequest.result !== ''
-    );
+    return isRequestSet(this.getThemeRequest);
   }
 
   @computed get hasLoadedCurrentTheme(): boolean {
-    return (
-      this.getThemeRequest.wasExecuted && this.getThemeRequest.result !== null
+    return hasLoadedRequest(this.getThemeRequest);
+  }
+
+  @computed get currentNumberFormat(): string {
+    return requestGetter(
+      this.getProfileNumberFormatRequest,
+      this.systemNumberFormat
     );
+  }
+
+  @computed get currentDateFormat(): string {
+    return this.currentLocale === 'en-US'
+      ? this.currentDateEnglishFormat
+      : this.currentDateJapaneseFormat;
+  }
+
+  @computed get currentDateEnglishFormat(): string {
+    return requestGetter(
+      this.getProfileDateFormatEnglishRequest,
+      this.systemDateFormatEnglish
+    );
+  }
+
+  @computed get currentDateJapaneseFormat(): string {
+    return requestGetter(
+      this.getProfileDateFormatJapaneseRequest,
+      this.systemDateFormatJapanese
+    );
+  }
+
+  @computed get currentTimeFormat(): string {
+    return requestGetter(
+      this.getProfileTimeFormatRequest,
+      this.systemTimeFormat
+    );
+  }
+
+  @computed get currentTimeFormatShort(): string {
+    return this.currentTimeFormat.replace(':ss', '');
   }
 
   @computed get termsOfUse(): string {
@@ -189,9 +262,30 @@ export default class ProfileStore extends Store {
     this._onReceiveSystemLocale(await detectSystemLocaleChannel.request());
   };
 
-  _updateLocale = async ({ locale }: { locale: string }) => {
-    await this.setProfileLocaleRequest.execute(locale);
-    await this.getProfileLocaleRequest.execute();
+  _finishInitialScreenSettings = action(() => {
+    this._consolidateUserSettings();
+    this.isInitialScreen = false;
+  });
+
+  _consolidateUserSettings = () => {
+    PROFILE_SETTINGS.forEach((param: string) => {
+      this._updateUserLocalSetting({ param });
+    });
+  };
+
+  _updateUserLocalSetting = async ({
+    param,
+    value,
+  }: {
+    param: string,
+    value?: string,
+  }) => {
+    // In case `value` is missing, it consolidates in the localstorage the default value
+    const consolidatedValue =
+      value || (this: any)[camelCase(['current', param])];
+    const { set, get } = getRequestKeys(param, this.currentLocale);
+    await (this: any)[set].execute(consolidatedValue);
+    await (this: any)[get].execute();
   };
 
   _updateTheme = async ({ theme }: { theme: string }) => {
@@ -221,10 +315,16 @@ export default class ProfileStore extends Store {
     this.getDataLayerMigrationAcceptanceRequest.execute();
   };
 
-  _redirectToLanguageSelectionIfNoLocaleSet = () => {
-    if (this.hasLoadedCurrentLocale && !this.isCurrentLocaleSet) {
+  _redirectToInitialSettingsIfNoLocaleSet = () => {
+    if (
+      (this.hasLoadedCurrentLocale && !this.isCurrentLocaleSet) ||
+      this.isInitialScreen
+    ) {
+      runInAction('Set `isInitialScreen` true', () => {
+        this.isInitialScreen = true;
+      });
       this.actions.router.goToRoute.trigger({
-        route: ROUTES.PROFILE.LANGUAGE_SELECTION,
+        route: ROUTES.PROFILE.INITIAL_SETTINGS,
       });
     }
   };
@@ -232,7 +332,11 @@ export default class ProfileStore extends Store {
   _redirectToTermsOfUseScreenIfTermsNotAccepted = () => {
     const termsOfUseNotAccepted =
       this.hasLoadedTermsOfUseAcceptance && !this.areTermsOfUseAccepted;
-    if (this.isCurrentLocaleSet && termsOfUseNotAccepted) {
+    if (
+      !this.isInitialScreen &&
+      this.isCurrentLocaleSet &&
+      termsOfUseNotAccepted
+    ) {
       this.actions.router.goToRoute.trigger({
         route: ROUTES.PROFILE.TERMS_OF_USE,
       });