+++ storybook/stories/transactions/TransactionsList.stories.js
@@ -3,7 +3,7 @@
 // @flow
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { boolean, number } from '@storybook/addon-knobs';
+import { boolean, number, select } from '@storybook/addon-knobs';
 import BigNumber from 'bignumber.js';
 import moment from 'moment';
 import faker from 'faker';
@@ -16,6 +16,11 @@ import {
   transactionTypes,
 } from '../../../source/renderer/app/domains/WalletTransaction';
 import WalletsWrapper from '../wallets/utils/WalletsWrapper';
+import { timeOptions, dateOptions } from '../_support/profileSettings';
+import {
+  DATE_ENGLISH_OPTIONS,
+  TIME_OPTIONS,
+} from '../../../source/renderer/app/config/profileConfig';
 
 // Screens
 import WalletTransactionsList from '../../../source/renderer/app/components/wallet/transactions/WalletTransactionsList';
@@ -29,6 +34,16 @@ storiesOf('Transactions|Transactions', module)
 
   .add('Transactions List', () => (
     <WalletTransactionsList
+      currentDateFormat={select(
+        'currentDateFormat',
+        dateOptions,
+        DATE_ENGLISH_OPTIONS[0].value
+      )}
+      currentTimeFormat={select(
+        'currentTimeFormat',
+        timeOptions,
+        TIME_OPTIONS[0].value
+      )}
       transactions={[
         ...Array.from(Array(number('Transactions Sent', 1))).map((x, i) =>
           generateTransaction(